# src/main/resources/api.yml
openapi: 3.0.0
info:
  title: Loyalty Card
  description: API для управления клиентами, картами лояльности, транзакциями и администраторами.
  version: 1.0.0
servers:
  - url: http://localhost:8080

paths:
  /customers:
    post:
      summary: Создать нового клиента
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCustomer'
      responses:
        '201':
          description: Клиент успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCustomerResponse'

  /customers/{customerID}:
    get:
      summary: Получить информацию о клиенте
      parameters:
        - name: customerID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Информация о клиенте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
    put:
      summary: Обновить данные клиента
      parameters:
        - name: customerID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomer'
      responses:
        '200':
          description: Клиент обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateCustomerResponse'
    delete:
      summary: Удалить клиента
      parameters:
        - name: customerID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Клиент удален
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteCustomerResponse'

  /loyalty-cards:
    post:
      summary: Создать карту лояльности
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewLoyaltyCard'
      responses:
        '201':
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateLoyaltyCardResponse'

  /loyalty-cards/{cardID}:
    get:
      summary: Получить информацию о карте лояльности
      parameters:
        - name: cardID
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о карте лояльности
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyCard'

  /loyalty-cards/{cardID}/balance:
    put:
      summary: Обновить баланс карты
      parameters:
        - name: cardID
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBalance'
      responses:
        '200':
          description: Баланс успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateBalanceResponse'

  /transactions:
    get:
      summary: Получить историю транзакций
      parameters:
        - name: cardID
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: История транзакций
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionHistory'

  /admins/login:
    post:
      summary: Авторизация администратора
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLogin'
      responses:
        '200':
          description: Авторизация успешна
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLoginResponse'

  /admins/register:
    post:
      summary: Регистрация нового администратора
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminRegister'
      responses:
        '201':
          description: Администратор успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRegisterResponse'

components:
  schemas:
    NewCustomer:
      type: object
      properties:
        name:
          type: string
          example: Иван Иванов
        phone:
          type: string
          example: +1234567890
        email:
          type: string
          example: example@email.com
      required:
        - name
        - phone
        - email

    Customer:
      type: object
      properties:
        customerID:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string

    UpdateCustomer:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string

    CreateCustomerResponse:
      type: object
      properties:
        customerID:
          type: integer
        message:
          type: string

    UpdateCustomerResponse:
      type: object
      properties:
        message:
          type: string

    DeleteCustomerResponse:
      type: object
      properties:
        message:
          type: string

    NewLoyaltyCard:
      type: object
      properties:
        customerID:
          type: integer
      required:
        - customerID

    CreateLoyaltyCardResponse:
      type: object
      properties:
        cardID:
          type: integer
        QRCode:
          type: string
        message:
          type: string

    LoyaltyCard:
      type: object
      properties:
        cardID:
          type: integer
        customerID:
          type: string
        QRCode:
          type: string
        balance:
          type: number
          format: float

    UpdateBalance:
      type: object
      properties:
        amount:
          type: number
          format: float
          example: 100.5
        operation:
          type: string
          enum: [add, subtract]
          example: add
        reason:
          type: string
          example: Награда за покупку

    UpdateBalanceResponse:
      type: object
      properties:
        message:
          type: string
        newBalance:
          type: number
          format: float

    TransactionHistory:
      type: array
      items:
        type: object
        properties:
          transactionID:
            type: integer
          cardID:
            type: integer
          amount:
            type: number
            format: float
          reason:
            type: string
          date:
            type: string
            format: date-time

    AdminLogin:
      type: object
      properties:
        username:
          type: string
        password:
          type: string

    AdminLoginResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
          enum: [manager, employee]

    AdminRegister:
      type: object
      properties:
        name:
          type: string
          example: Иван Иванов
        email:
          type: string
          format: email
          example: admin@example.com
        password:
          type: string
          format: password
          example: StrongPass123!
      required:
        - name
        - email
        - password

    AdminRegisterResponse:
      type: object
      properties:
        adminID:
          type: integer
        message:
          type: string